simplify xp model
