+++ libs/common/src/models/index.ts
@@ -3,5 +3,5 @@ export * from './event.model';
 export * from './github-shared.model';
 export * from './user-activity.model';
 export * from './user.model';
-export * from './xp.model';
-export * from './xp-history.model';
+export * from './xp-domain.model';
+
