+++ libs/common/src/models/user.model.ts
@@ -5,15 +5,14 @@ export interface CreateUserInput {
   email: string | null;
   displayName: string | null;
   photoUrl: string | null;
-  // githubUsername?: string;
 }
 
-// User Document
+// later TODO: User could additionaly extend UserXpData
 export interface User extends BaseDocument {
   email: string;
   displayName: string;
   photoUrl: string;
   lastLogin: string;
   active: boolean;
-  uid?: string; // via Firebase Auth
+  uid?: string;
 }
